newuser/pyproject.toml
[tool.poetry]
name = "newuser"
version = "0.1.0"
description = ""
authors = ["Eric Preston <eric.preston@redis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^4.5.5"
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
pytest-env = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
